requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2018.4.16
chardet==3.0.4
dbus-python==1.2.8
idna==2.7
lxml==4.2.3
pkg-resources==0.0.0
requests==2.19.1
urllib3==1.23
